Harden SMS operations queue monitoring
